Some initial proposed changes to the script file format
Some of this is actual format changes, other parts would just be
considered best practices.

I don't want to require more layers then needed.  A transform can be
the only thing in the file, or a collection of transforms.

I think we can support the collection of transforms in the script
namespace whether or not the scripting engine is goint to be separating
them.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexisrisk.com>